% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08_schema.R
\name{check_schema}
\alias{check_schema}
\title{Verifie qu'un data.frame respecte un me_schema}
\usage{
check_schema(data, schema)
}
\arguments{
\item{data}{Un data.frame.}

\item{schema}{Un \code{me_schema}.}
}
\value{
\code{TRUE} invisible si conforme ; leve \code{me_validation_error} sinon.
}
\description{
Verifie qu'un data.frame respecte un me_schema
}
