% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_generics.R
\name{compute_indicators}
\alias{compute_indicators}
\title{Calculer les indicateurs d'un objet}
\usage{
compute_indicators(x, ...)
}
\arguments{
\item{x}{Un objet MEverse portant des indicateurs (ex. Project).}

\item{...}{Arguments additionnels passes aux methodes.}
}
\description{
Calculer les indicateurs d'un objet
}
