% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_generics.R
\name{describe}
\alias{describe}
\title{Decrire un objet MEverse en texte structure (voir idee IA-ready)}
\usage{
describe(x, ...)
}
\arguments{
\item{x}{Un objet MEverse (Indicator, Dataset, Project, ...).}

\item{...}{Arguments additionnels passes aux methodes.}
}
\description{
Decrire un objet MEverse en texte structure (voir idee IA-ready)
}
