% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_generics.R
\name{generate_report}
\alias{generate_report}
\title{Generer un rapport a partir d'un objet}
\usage{
generate_report(x, ...)
}
\arguments{
\item{x}{Un objet MEverse a partir duquel generer un rapport (ex. Project).}

\item{...}{Arguments additionnels passes aux methodes.}
}
\description{
Generer un rapport a partir d'un objet
}
