% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_dataset.R
\name{me_dataset}
\alias{me_dataset}
\title{Jeu de donnees MEverse}
\usage{
me_dataset(
  name = character(0),
  data = (function (.data = list(), row.names = NULL) 
 {
     if (is.null(row.names)) {

            list2DF(.data)
     }
     else {
         out <- list2DF(.data,
    length(row.names))
         attr(out, "row.names") <- row.names
         out
     }

    })(),
  metadata = me_metadata()
)
}
\arguments{
\item{name}{Nom court du dataset (ex. "baseline_2026").}

\item{data}{Un \code{data.frame}.}

\item{metadata}{Un objet \code{me_metadata}.}
}
\description{
Conteneur pour des donnees brutes ou nettoyees. \code{mecore} ne definit
que la structure ; la logique de nettoyage/import appartient a \code{medata}.
}
