% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09_indicator_registry.R
\name{me_indicator_from_sdg}
\alias{me_indicator_from_sdg}
\title{Instancier un me_indicator a partir d'un code ODD standard}
\usage{
me_indicator_from_sdg(code, datasets, formula, value = NULL)
}
\arguments{
\item{code}{Code de l'indicateur standard (ex. "4.1.1").}

\item{datasets}{Liste de \code{me_dataset} sources.}

\item{formula}{Formule de calcul specifique au contexte du projet.}

\item{value}{Valeur calculee (optionnelle a la creation).}
}
\description{
Instancier un me_indicator a partir d'un code ODD standard
}
