% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_logframe.R
\name{me_logframe}
\alias{me_logframe}
\title{Cadre logique MEverse}
\usage{
me_logframe(goal = character(0), results = list(), hypotheses = character(0))
}
\arguments{
\item{goal}{Objectif general.}

\item{results}{Liste de \code{me_result}.}

\item{hypotheses}{Liste de chaines de caracteres (hypotheses/risques).}
}
\description{
Un \code{me_project} a au plus UN \code{me_logframe} actif (cardinalite 0..1).
}
