% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_project.R
\name{me_project}
\alias{me_project}
\title{Projet MEverse}
\usage{
me_project(
  name = character(0),
  metadata = me_metadata(),
  datasets = list(),
  indicators = list(),
  logframe = me_logframe()
)
}
\arguments{
\item{name}{Nom du projet.}

\item{metadata}{Un objet \code{me_metadata}.}

\item{datasets}{Liste de \code{me_dataset} (peut etre vide).}

\item{indicators}{Liste de \code{me_indicator} (peut etre vide).}

\item{logframe}{\code{me_logframe} ou \code{NULL}.}
}
\description{
Racine agregeante. Un \code{me_project} a exactement une \code{me_metadata},
zero ou plusieurs \code{me_dataset}, zero ou plusieurs \code{me_indicator},
au plus un \code{me_logframe}. Voir ARCHITECTURE.md §2.2.
}
