% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_logframe.R
\name{me_result}
\alias{me_result}
\title{Resultat d'un cadre logique}
\usage{
me_result(label = character(0), indicators = list())
}
\arguments{
\item{label}{Libelle du resultat (ex. "R1 : acces ameliore aux services").}

\item{indicators}{Liste de \code{me_indicator} associes (peut etre vide en construction).}
}
\description{
Resultat d'un cadre logique
}
