% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_generics.R
\name{me_validate}
\alias{me_validate}
\title{Valider un objet MEverse}
\usage{
me_validate(x, ...)
}
\arguments{
\item{x}{Un objet MEverse (Project, Dataset, Indicator, ...).}

\item{...}{Arguments additionnels passes aux methodes.}
}
\description{
Retourne l'objet inchange si valide (pour permettre le chainage avec
le pipe natif \verb{|>}) ; leve une condition \code{me_validation_error} sinon.
}
