% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_conditions.R
\name{me_validation_error}
\alias{me_validation_error}
\title{Lever une erreur de validation metier MEverse}
\usage{
me_validation_error(message, ...)
}
\arguments{
\item{message}{Message d'erreur affiche a l'utilisateur.}

\item{...}{Donnees additionnelles attachees a la condition.}
}
\description{
Lever une erreur de validation metier MEverse
}
