% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/12_pipeline.R
\name{run_pipeline}
\alias{run_pipeline}
\title{Executer un pipeline sur un objet, avec journalisation}
\usage{
run_pipeline(pipeline, x)
}
\arguments{
\item{pipeline}{Un \code{me_pipeline}.}

\item{x}{Objet de depart (ex. un \code{me_project}).}
}
\value{
Une liste \code{list(result = ..., pipeline = ...)} ; \code{pipeline}
contient le journal d'execution mis a jour (horodatage, nom d'etape).
}
\description{
Executer un pipeline sur un objet, avec journalisation
}
